+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Maximum stack trace lines per thread is now 30 by default.
     - Maximum stack trace line length is now 200 by default. 
 - ! Minor breaking change ! After initialization, the logging/debugging mode can only be changed with `Countly.setLoggingEnabled` instead of `Countly.debug` now.
+- When recording internal events with 'add_event', the respective feature consent will now be checked instead of just the 'events' consent. 
 - Fixed a bug where the SDK throws a `Bulk user storage exception` due to a missing folder
 - Increased the default max event batch size to 100.
 - Logs are now color coded and indicate log levels.